New structure for allowing shared updates.

+++ scripts/after_base_fetch.sh
@@ -13,7 +13,9 @@ if [ ! -f ${PROJECT_ROOT_PATH}/src/pages/index.vue ]; then
 fi
 
 # link to self
-cp ./${S_BASE_NAME}/templates/${S_BASE_NAME}.sh ${PROJECT_ROOT_PATH}
+if [ ! -f ${PROJECT_ROOT_PATH}/${S_BASE_NAME}.sh ]; then
+  cp ./${S_BASE_NAME}/templates/${S_BASE_NAME}.sh ${PROJECT_ROOT_PATH}
+fi
 
 DATE=`date '+%Y-%m-%d_%H%M%S'`
 STASH_DIR=_stash_${DATE}
